requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
opencv-python-headless==4.9.0.80
pandas==2.2.0
tqdm==4.66.1
plotly==5.18.0
gunicorn==21.2.0
python-dotenv==1.0.1
